pyproject.toml
[project]
name = "scout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.132.0",
    "feedparser>=6.0.12",
    "jinja2>=3.1.6",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pypdf>=6.7.2",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]
